+++ CHANGELOG.md
@@ -10,6 +10,11 @@
 - feat: timebar supports afunction getValue  to returns the value (for trend line of trend timebar) according to each node or edge by user;
 - feat: timebar supports to configure a boolean changeData to control the filter way, true means filters by graph[dot]changeData, false means filters by graph[dot]showItem and graph[dot]hideItem;
 - feat: timebar supports to configure a function shouldIgnore to return true or false by user to decide whether the node or the edge should be ignored while filtering;
+- feat: translate graph with animation;
+- feat: zoom graph with animation;
+- fix: edge label padding bug, closes: #3346;
+- fix: update node with iconfont icon, the icon is updated to a wrong position, closes: #3348;
+
 
 #### 4.5.0